backend/requirements.txt
fastapi[all]
langchain
langchain-core
langchain-openai
langchain-google-genai
uvicorn
python-dotenv
langchain-community
